+++ docs/changelog.rst
@@ -12,6 +12,7 @@ Added
 ^^^^^
 - New skill socres: KGE 2012, KGE non-parametric (2018), KGE non-parametric flood (2018).
 - new rasterio inverse distance weighting method ("rio_idw") in raster.interpolate_na
+- Add {model}, {scenario}, {realization} as placeholders of the path of a data catalog key.
 
 Changed
 ^^^^^^^